Updated on 2026-08-27

+++ features/token-recieve/impl/src/main/java/com/tangem/features/tokenreceive/ui/state/QrCodeUM.kt
@@ -3,7 +3,7 @@ package com.tangem.features.tokenreceive.ui.state
 import com.tangem.core.ui.extensions.TextReference
 
 internal data class QrCodeUM(
-    val onCopyClick: (Int) -> Unit,
+    val onCopyClick: () -> Unit,
     val onShareClick: (String) -> Unit,
     val addressName: TextReference,
     val addressValue: String,
